feat: render external and internal links
