+++ app/articles/[slug]/page.tsx
@@ -5,14 +5,12 @@ import {
 	fetchTwoNewestArticlesBySlug,
 	sanityClient,
 } from "@/sanity/api";
-import PortableText from "react-portable-text";
-import { Box, Table, Typography } from "@mui/material";
+import { Box, Typography } from "@mui/material";
 import { BackIcon } from "@/app/icons/BackIcon";
 import Link from "next/link";
-import Image from "next/image";
 import imageUrlBuilder from "@sanity/image-url";
-import { getImageDimensions } from "@sanity/asset-utils";
 import ArticleCard from "@/app/components/ArticleCard";
+import PortableTextStyled from "@/app/components/PortableTextStyled";
 
 export default async function PostPage({
 	params,
@@ -39,63 +37,7 @@ export default async function PostPage({
 				<Typography className={"text-base"}> Back to articles </Typography>
 			</Link>
 			{article.body !== null ? (
-				<PortableText
-					dataset={process.env.NEXT_PUBLIC_SANITY_DATASET}
-					projectId={process.env.NEXT_PUBLIC_SANITY_PROJECT_ID}
-					content={article.body}
-					serializers={{
-						h1: (props: any) => (
-							<h1 className="text-5xl xs:text-6xl mb-12" {...props} />
-						),
-						h2: (props: any) => (
-							<h2 className="text-4xl xs:text-5xl mb-6" {...props} />
-						),
-						h3: (props: any) => (
-							<h3 className="text-3xl xs:text-4xl mb-6" {...props} />
-						),
-						h4: (props: any) => (
-							<h4 className="text-2xl xs:text-3xl mb-6" {...props} />
-						),
-						normal: (props: any) => (
-							<p className="text-base mb-10" {...props} />
-						),
-						table: (props: any) => <Table {...props} />, //Dette funker ikke
-						li: ({ children }: any) => (
-							<li className="ml-4 text-base list-disc">{children}</li>
-						),
-						ul: ({ children }: any) => (
-							<ul className={"ml-4 mb-6"}>{children}</ul>
-						),
-						blockquote: ({ children }: any) => (
-							<blockquote className={"ml-4 text-base italic"}>
-								{children}
-							</blockquote>
-						),
-						link: ({ href, children }: any) => (
-							<a
-								href={href}
-								className="text-primary-main underline hover:no-underline"
-							>
-								{children}
-							</a>
-						),
-						image: (value: any, props: any) => {
-							const { width, height } = getImageDimensions(value);
-
-							return (
-								<Image
-									src={value !== null ? builder.image(value).toString() : ""}
-									alt={"Article photo"}
-									loading={"lazy"}
-									height={height}
-									width={width}
-									className={"w-auto h-auto my-20"}
-									{...props}
-								/>
-							);
-						},
-					}}
-				/>
+				<PortableTextStyled content={article.body} />
 			) : (
 				<Typography>No content published yet</Typography>
 			)}
