+++ app/resources/[slug]/page.tsx
@@ -1,15 +1,11 @@
 import {
 	fetchDocumentBySlug,
 	fetchDocuments,
-	sanityClient,
 } from "@/sanity/api";
-import PortableText from "react-portable-text";
-import { Box, Table, Typography } from "@mui/material";
+import PortableTextStyled from "@/app/components/PortableTextStyled";
+import { Box, Typography } from "@mui/material";
 import { BackIcon } from "@/app/icons/BackIcon";
 import Link from "next/link";
-import Image from "next/image";
-import imageUrlBuilder from "@sanity/image-url";
-import { getImageDimensions } from "@sanity/asset-utils";
 import CodeBlock from "@/app/components/CodeBlock";
 
 export default async function DocumentPage({
@@ -18,8 +14,6 @@ export default async function DocumentPage({
 	params: { slug: string };
 }) {
 	const document = await fetchDocumentBySlug(params.slug);
-	const builder = imageUrlBuilder(sanityClient);
-
 	return (
 		<Box
 			className={
@@ -36,62 +30,9 @@ export default async function DocumentPage({
 				<Typography className={"text-base"}> Back to resources </Typography>
 			</Link>
 			{document.body !== null && document.body !== undefined ? (
-				<PortableText
-					dataset={process.env.NEXT_PUBLIC_SANITY_DATASET}
-					projectId={process.env.NEXT_PUBLIC_SANITY_PROJECT_ID}
-					content={document.body ?? []}
-					serializers={{
-						h1: (props: any) => (
-							<h1 className="text-5xl xs:text-6xl mb-12" {...props} />
-						),
-						h2: (props: any) => (
-							<h2 className="text-4xl xs:text-5xl mb-6" {...props} />
-						),
-						h3: (props: any) => (
-							<h3 className="text-3xl xs:text-4xl mb-6" {...props} />
-						),
-						h4: (props: any) => (
-							<h4 className="text-2xl xs:text-3xl mb-6" {...props} />
-						),
-						normal: (props: any) => (
-							<p className="text-base mb-10" {...props} />
-						),
-						table: (props: any) => <Table {...props} />,
-						li: ({ children }: any) => (
-							<li className="ml-4 text-base list-disc">{children}</li>
-						),
-						ul: ({ children }: any) => (
-							<ul className={"ml-4 mb-6"}>{children}</ul>
-						),
-						blockquote: ({ children }: any) => (
-							<blockquote className={"ml-4 text-base italic"}>
-								{children}
-							</blockquote>
-						),
-						link: ({ href, children }: any) => (
-							<a
-								href={href}
-								className="text-primary-main underline hover:no-underline"
-							>
-								{children}
-							</a>
-						),
-						image: (value: any, props: any) => {
-							const { width, height } = getImageDimensions(value);
-
-							return (
-								<Image
-									src={value !== null ? builder.image(value).toString() : ""}
-									alt={"Article photo"}
-									loading={"lazy"}
-									height={height}
-									width={width}
-									className={"w-auto h-auto my-20"}
-									{...props}
-								/>
-							);
-						},
-					}}
+				<PortableTextStyled
+					content={document.body}
+					internalLinkPrefix={"/resources"}
 				/>
 			) : (
 				<Typography>No content published yet</Typography>
